Renaming services module + misc cleanup
