+++ src/fides/api/api/v1/endpoints/drp_endpoints.py
@@ -43,10 +43,10 @@ from fides.common.api import scope_registry as scopes
 from fides.common.api.v1 import urn_registry as urls
 from fides.config import CONFIG
 from fides.config.config_proxy import ConfigProxy
-from fides.services.messaging.messaging_service import (
+from fides.service.messaging.messaging_service import (
     check_and_dispatch_error_notifications,
 )
-from fides.services.privacy_request.privacy_request_service import queue_privacy_request
+from fides.service.privacy_request.privacy_request_service import queue_privacy_request
 
 router = APIRouter(tags=["DRP"], prefix=urls.V1_URL_PREFIX)
