+++ src/fides/api/api/v1/exception_handlers.py
@@ -4,13 +4,13 @@ from fastapi import Request
 from fastapi.responses import JSONResponse
 from starlette.status import HTTP_500_INTERNAL_SERVER_ERROR
 
-from fides.api.common_exceptions import FunctionalityNotConfigured
+from fides.api.common_exceptions import RedisNotConfigured
 
 
 class ExceptionHandlers:
     @staticmethod
-    def functionality_not_configured_handler(
-        request: Request, exc: FunctionalityNotConfigured
+    def redis_not_configured_handler(
+        request: Request, exc: RedisNotConfigured
     ) -> JSONResponse:
         return JSONResponse(
             status_code=HTTP_500_INTERNAL_SERVER_ERROR, content={"message": str(exc)}
@@ -19,5 +19,5 @@ class ExceptionHandlers:
     @classmethod
     def get_handlers(
         cls,
-    ) -> List[Callable[[Request, FunctionalityNotConfigured], JSONResponse]]:
-        return [ExceptionHandlers.functionality_not_configured_handler]
+    ) -> List[Callable[[Request, RedisNotConfigured], JSONResponse]]:
+        return [ExceptionHandlers.redis_not_configured_handler]
