+++ src/fides/api/util/connection_util.py
@@ -59,7 +59,7 @@ from fides.api.service.connectors.saas.connector_registry_service import (
 )
 from fides.api.util.logger import Pii
 from fides.common.api.v1.urn_registry import CONNECTION_TYPES, SAAS_CONFIG
-from fides.services.privacy_request.privacy_request_service import queue_privacy_request
+from fides.service.privacy_request.privacy_request_service import queue_privacy_request
 
 # pylint: disable=too-many-nested-blocks,too-many-branches,too-many-statements
