+++ tests/ops/api/v1/endpoints/test_consent_request_endpoints.py
@@ -246,7 +246,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request(self, mock_dispatch_message, api_client, url):
         data = {"identity": {"email": "test@example.com"}}
         response = api_client.post(url, json=data)
@@ -258,7 +258,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_identity_present(
         self,
         mock_dispatch_message,
@@ -288,7 +288,7 @@ class TestConsentRequest:
         "messaging_config",
         "sovrn_email_connection_config",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_subject_verification_disabled_no_email(
         self, mock_dispatch_message, api_client, url
     ):
@@ -302,7 +302,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_phone_number(self, mock_dispatch_message, api_client, url):
         data = {"identity": {"phone_number": "+3368675309"}}
         response = api_client.post(url, json=data)
@@ -314,7 +314,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_email_and_phone_use_config(
         self,
         mock_dispatch_message,
@@ -341,7 +341,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_email_and_phone_default_to_email(
         self,
         mock_dispatch_message,
@@ -372,7 +372,7 @@ class TestConsentRequest:
         "messaging_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_with_source(
         self,
         mock_dispatch_message,
@@ -399,7 +399,7 @@ class TestConsentRequest:
         "sovrn_email_connection_config",
         "subject_identity_verification_required",
     )
-    @patch("fides.services.messaging.messaging_service.dispatch_message")
+    @patch("fides.service.messaging.messaging_service.dispatch_message")
     def test_consent_request_with_custom_privacy_request_fields(
         self,
         mock_dispatch_message,
